shared/mcp-servers/prm-graphapi/requirements.txt
# FastMCP and MCP dependencies
fastmcp==2.13.0.2
mcp==1.19.0

# Web framework
starlette==0.49.1
uvicorn[standard]==0.38.0

# Azure and Microsoft Graph
azure-identity>=1.15.0
azure-core>=1.29.0
msgraph-sdk>=1.2.0
msal>=1.26.0

# Utilities
python-dotenv>=1.0.0
